pyproject.toml
[project]
name = "sensei-web"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "allpairspy>=2.5.1",
    "boto3>=1.38.30",
    "chatbot-tracer>=0.2.7",
    "cryptography>=45.0.3",
    "django>=5.2.2",
    "django-cors-headers>=4.7.0",
    "django-rest-knox>=5.0.2",
    "djangorestframework>=3.16.0",
    "dotenv>=0.9.9",
    "jinja2>=3.1.6",
    "langchain>=0.3.25",
    "openai>=1.84.0",
    "pandas>=2.3.0",
    "pillow>=11.2.1",
    "psutil>=7.0.0",
    "pydantic>=2.11.5",
    "pygame>=2.6.1",
    "pymupdf>=1.26.1",
    "pyyaml>=6.0.2",
    "scikit-learn>=1.6.1",
    "selenium>=4.33.0",
    "tiktoken>=0.9.0",
    "webdriver-manager>=4.0.2",
]

[build-system]
requires = ["setuptools>=61.0"] 
build-backend = "setuptools.build_meta"
backend-path = ["."] 

[tool.setuptools]
# This tells setuptools that the Python package(s) for 'sensei-web' are found in the 'backend' directory.
# Adjust if your main source directory for 'sensei-web' has a different name.
packages = ["backend"]

[tool.ruff]
# default is too small
line-length = 120
target-version = "py311"
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.lint]
# E: pycode style (error)
# W: pycode style (warning)
# F: Pyflakes
# I: ishort
# ANN: type annotations
# D: docstrings
# B: bugbear
select = [
    "ALL"
]
ignore = [
    "E501", # line too long
    # "D100", "D101", "D102", "D103", "D104", "D105", "D106", "D107", # Common missing docstring rules
    "COM812", # Creates incompatibility with formatter
    "T201", # Doesnt let you use prints
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
]

# Alternatively, for more complex scenarios or finer control (e.g., multiple top-level package directories for sensei-web):
# [tool.setuptools.packages.find]
# where = ["."]  # Look for packages in the project root ('.') or a specific directory like 'src'
# include = ["backend*"]  # Include 'backend' and any sub-packages (e.g., backend.api)
# exclude = ["frontend*", "tracer*", "tests*"] # Explicitly exclude other directories from this package
